microservices/chat-service/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
openai
redis
aioredis
httpx
fastapi-limiter
